Move root classes to a subnamespace and remove useless phpdoc
