+++ templates/content/homepage.html.twig
@@ -79,7 +79,7 @@
                 </div>
                 <div class="step-feature-text">
                     <h3>3. Inform Users</h3>
-                    <p>After <a href="https://github.com/jolicode/secret-santa/blob/master/src/Rudolph.php">Rudolph</a> shuffled the users, we throw at each participant a private message with his peer name and an optional admin message.</p>
+                    <p>After <a href="https://github.com/jolicode/secret-santa/blob/master/src/Santa/Rudolph.php">Rudolph</a> shuffled the users, we throw at each participant a private message with his peer name and an optional admin message.</p>
                 </div>
             </section>
